+++ app/assets/main.ts
@@ -22,6 +22,10 @@ app.use(CoreSprinkle)
 import AccountSprinkle from '@userfrosting/sprinkle-account'
 app.use(AccountSprinkle, { router })
 
+/** Setup Admin Sprinkle */
+import AdminSprinkle from '@userfrosting/sprinkle-admin'
+app.use(AdminSprinkle)
+
 /** Setup Theme */
 import '@userfrosting/theme-pink-cupcake/less/main.less'
 import PinkCupcake from '@userfrosting/theme-pink-cupcake'